requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
timm>=0.9.0
opencv-python>=4.8.0
albumentations>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
